move types and consts to seperate folders
