+++ src/components/__tests__/TaskForm.spec.ts
@@ -2,7 +2,7 @@ import { createPinia, setActivePinia } from 'pinia'
 import { beforeAll, describe, it, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
 import { useTaskStore } from '@/stores/task'
-import type { Task } from '@/stores/task'
+import type { Task } from '@/types'
 import TaskForm from '@/components/TaskForm.vue'
 
 describe('TaskForm', () => {
