+++ src/components/__tests__/TasksList.spec.ts
@@ -3,8 +3,8 @@ import { beforeAll, describe, it, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
 import { useTaskStore } from '@/stores/task'
 import { useListStore } from '@/stores/list'
-import type { List } from '@/stores/list'
-import type { Task } from '@/stores/task'
+import type { List } from '@/types'
+import type { Task } from '@/types'
 import TasksList from '@/components/TasksList/Index.vue'
 
 describe('TasksList', () => {
